Module UTop_complete

OCaml completion.

val complete : phrase_terminator:string -> input:string -> int * (string * string) list

complete ~phrase_terminator ~input returns the start of the completed word in input and the list of possible completions with their suffixes.

val reset : unit -> unit

Reset global cache. It must be called before each interactive read line.